requirements.txt
flask
flask-cors
pytesseract
pymupdf
Pillow
gunicorn
